[PATCH] code generation for EXPR_INVOKE when used with EXPR_ARG
We need to push the function return value in %eax to stack when EXPR_INVOKE
is used with EXPR_ARG.  Note: the acceptance test now hangs instead of
crashing.

+++ jit/x86-objcode.h
@@ -18,7 +18,7 @@ static inline void init_insn_sequence(struct insn_sequence *is, void *code,
 
 void x86_emit_prolog(struct insn_sequence *);
 void x86_emit_epilog(struct insn_sequence *);
-void x86_emit_push_imm32(struct insn_sequence *, int);
+void x86_emit_push_imm32(struct insn_sequence *, unsigned long);
 void x86_emit_call(struct insn_sequence *, void *);
 void x86_emit_ret(struct insn_sequence *);
 void x86_emit_add_imm8_reg(struct insn_sequence *, unsigned char, enum reg);
